% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandtridiag.R
\name{bandtridiag}
\alias{bandtridiag}
\title{Generate band matrix with three diagonals on the lower side}
\usage{
bandtridiag(lower, middle, main)
}
\arguments{
\item{lower}{vector of length p-1, the elements in the second diagonal below the main diagonal.}

\item{middle}{vector of length p-1, the elements in the first diagonal below the main diagonal.}

\item{main}{vector of length p-1, the elements in the main diagonal.}
}
\description{
Generate band matrix of dimention p*(p-2) that has nonzero elements only on the main diagonal, the first two diagonals below the main diagonal.
}
\examples{
lower<-1:3
middle<-4:6
main<-7:9
bandtridiag(lower, middle, main)

}
