% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_logistic.R
\name{curvature_logistic}
\alias{curvature_logistic}
\title{Square of curvature of logistic curve}
\usage{
curvature_logistic(theta1, theta2, theta3)
}
\arguments{
\item{theta1}{parameter in logistic function that defines the asymptote.}

\item{theta2}{parameter in logistic function.}

\item{theta3}{parameter in logistic function.}
}
\value{
function of square of curvature values.
}
\description{
Compute the square of curvature values of 3-parameter logistic function.
}
\details{
The 3-parameter logistic curve is given by \deqn{f(t)=\theta_1/(1+exp(\theta_2 t + \theta_3))} where \eqn{\theta_1} defines the asymptote, and \eqn{-\theta_3/\theta_2} defines the inflection point.
}
\examples{
theta1<-1
theta2<--10
theta3<-5
t<-seq(0,1,by=0.01)
drv2sq_fun<-curvature_logistic(theta1,theta2,theta3)
plot(t,drv2sq_fun(t),type="l")

}
