% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_logistic_pert.R
\name{curvature_logistic_pert}
\alias{curvature_logistic_pert}
\title{Square of curvature of logistic curve with perturbation}
\usage{
curvature_logistic_pert(theta1, theta2, theta3, mu, sigma)
}
\arguments{
\item{theta1}{parameter in logistic function that defines the asymptote.}

\item{theta2}{parameter in logistic function.}

\item{theta3}{parameter in logistic function.}

\item{mu}{parameter in Gaussian function.}

\item{sigma}{parameter in Gaussian function.}
}
\value{
function of square of curvature values.
}
\description{
Compute the square of curvature values of a mixture of 3-parameter logistic function and a Gaussian function.
}
\details{
The curve is given by \deqn{f(t)=\theta_1/(1+exp(\theta_2 t + \theta_3)) - 0.02/\sigma * exp(-(t-\mu)^2/(2\sigma^2))}, where \eqn{\theta_1} defines the asymptote, and \eqn{-\theta_3/\theta_2} defines the inflection point.
}
\examples{
theta1<-1
theta2<--10
theta3<-5
mu<-0.7
sigma<-0.002
t<-seq(0,1,by=0.01)
drv2sq_fun<-curvature_logistic_pert(theta1,theta2,theta3,mu,sigma)
plot(t,drv2sq_fun(t),type="l")

}
