% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency}
\alias{efficiency}
\title{efficiency}
\usage{
efficiency(t1, t2, eta = 1, curvature)
}
\arguments{
\item{t1}{A vector of time points, usually the optimal design points.}

\item{t2}{A vector of time points, usually the uniform design points.}

\item{eta}{smoothing parameter.}

\item{curvature}{square of prior curvature, function of t.}
}
\value{
A vector of optimal design points.
}
\description{
Compute the D-efficiency between two designs.
}
\examples{
n<-4
drv2sq_fun<-curvature_logistic(theta1=1,theta2=-10,theta3=5)
OD_t<-ODsplines(n=n,curvature=drv2sq_fun)
t_u<-seq(0,1,length.out = n)
efficiency(OD_t,t_u,curvature=drv2sq_fun)

}
