% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODsplines.R
\name{ODsplines}
\alias{ODsplines}
\title{optimal design for single population}
\usage{
ODsplines(n, eta = 1, curvature, interv = rep(0.001, n - 1),
  seed = 1)
}
\arguments{
\item{n}{number of design points.}

\item{eta}{smoothing parameter.}

\item{curvature}{square of prior curvature, function of t.}

\item{interv}{a vector of length n-1, the mininal intervals between consecutive design points.}

\item{seed}{an integer.}
}
\value{
A vector of optimal design points.
}
\description{
Derive the optimal design points of the adaptive smoothing splines. The prior knowledge is given  by the square of curvature values from a single population.
}
\examples{
n<-4
OD_t<-ODsplines(n=n,eta=1,
curvature=curvature_logistic(theta1=1,theta2=-10,theta3=5))

}
