% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridiag.R
\name{tridiag}
\alias{tridiag}
\title{Generate tridiagonal matrix}
\usage{
tridiag(upper, lower, main)
}
\arguments{
\item{upper}{vector of length p-1, the elements in the first diagonal above the main diagonal.}

\item{lower}{vector of length p-1, the elements in the first diagonal below the main diagonal.}

\item{main}{vector of length p,the elements in the main diagonal.}
}
\description{
Generate square tridiagonal matrix of dimention p that has nonzero elements only on the main diagonal, the first diagonal below and above the main diagonal.
}
\examples{
main<-1:4
upper<-5:7
lower<-8:10
tridiag(upper, lower, main)

}
